redoing assignment from start

+++ routes/api/product-routes.js
@@ -1,5 +1,4 @@
 const router = require('express').Router();
-const sequelize = require("../../config/connection");
 const { Product, Category, Tag, ProductTag } = require('../../models');
 
 // The `/api/products` endpoint
@@ -7,52 +6,13 @@ const { Product, Category, Tag, ProductTag } = require('../../models');
 // get all products
 router.get('/', (req, res) => {
   // find all products
-  Product.findAll({
   // be sure to include its associated Category and Tag data
-  include: [
-    Category,
-    {
-      model: Tag,
-      through: ProductTag,
-      as: "tags",
-    },
-  ],
-})
-.then((dbProductData) => res.json(dbProductData))
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json(err);
-    });
 });
 
 // get one product
 router.get('/:id', (req, res) => {
   // find a single product by its `id`
   // be sure to include its associated Category and Tag data
-  Product.findOne({
-    where: {
-      id: req.params.id,
-    },
-    include: [
-      Category,
-      {
-        model: Tag,
-        through: ProductTag,
-        as: "tags",
-      },
-    ],
-  })
-  .then((dbProductData) => {
-    if (!dbProductData) {
-      res.status(404).json({ message: "No post found with this id" });
-      return;
-    }
-    res.json(dbProductData);
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json(err);
-    });
 });
 
 // create new product
@@ -65,12 +25,7 @@ router.post('/', (req, res) => {
       tagIds: [1, 2, 3, 4]
     }
   */
-    Product.create({
-      product_name: req.body.product_name,
-      price: req.body.price,
-      stock: req.body.stock,
-      tagIds: req.body.tagIds,
-    })
+  Product.create(req.body)
     .then((product) => {
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
       if (req.body.tagIds.length) {
@@ -101,51 +56,44 @@ router.put('/:id', (req, res) => {
     },
   })
     .then((product) => {
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      const newProductTags = req.body.tagIds
-      .filter((tag_id) => !productTagIds.includes(tag_id))
-      .map((tag_id) => {
-        return {
-          product_id: req.params.id,
-          tag_id,
-        };
-      });
+      if (req.body.tagIds && req.body.tagIds.length) {
         
-      const productTagsToRemove = productTags
-      .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-      .map(({ id }) => id);
+        ProductTag.findAll({
+          where: { product_id: req.params.id }
+        }).then((productTags) => {
+          // create filtered list of new tag_ids
+          const productTagIds = productTags.map(({ tag_id }) => tag_id);
+          const newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            };
+          });
 
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+            // figure out which ones to remove
+          const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
+                  // run both actions
+          return Promise.all([
+            ProductTag.destroy({ where: { id: productTagsToRemove } }),
+            ProductTag.bulkCreate(newProductTags),
+          ]);
+        });
+      }
+
+      return res.json(product);
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
     .catch((err) => {
+      // console.log(err);
       res.status(400).json(err);
     });
 });
-router.delete("/:id", (req, res) => {
-  Product.destroy({
-    where: {
-      id: req.params.id,
-    },
-  })
-  .then((dbProductData) => {
-    if (!dbProductData) {
-      res.status(404).json({ message: "No product found with this id" });
-      return;
-    }
 
-      res.json(dbProductData);
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json(err);
-    });
+router.delete('/:id', (req, res) => {
+  // delete one product by its `id` value
 });
 
 module.exports = router;
